segmentation/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Segmentation models
segmentation-models-pytorch>=0.3.3

# Data processing
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Visualization & progress
matplotlib>=3.7.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
